+++ cardGame/cardGame/src/components/LevelContainer.jsx
@@ -1,0 +1,58 @@
+import styled from "styled-components";
+import { Level } from "../data/level";
+
+const LevelContainer = () => {
+  // 레벨 버튼 클릭 시 카드 수 변경
+  const ClickedLv = (e) => {
+    setScore(0);
+    setNowLevel(e.target.value);
+    switch (e.target.value) {
+      case "EASY":
+        return setCardsNum(EasyVersion);
+      case "NORMAL":
+        return setCardsNum(NormalVersion);
+      case "HARD":
+        return setCardsNum(HardVersion);
+    }
+  };
+  return (
+    <StLevenContainer>
+      {Level.map((level) => {
+        return (
+          <StLevelBtn
+            key={level.id}
+            value={level.lv}
+            type="button"
+            className={"btn" + (level.lv == nowLevel ? " active" : "")}
+            onClick={ClickedLv}
+          >
+            {level.lv}
+          </StLevelBtn>
+        );
+      })}
+    </StLevenContainer>
+  );
+};
+
+export default LevelContainer;
+
+const StLevenContainer = styled.nav`
+  display: flex;
+  justify-content: center;
+`;
+
+const StLevelBtn = styled.button`
+  width: 5rem;
+  margin: 1rem 1rem;
+  background-color: ${({ theme }) => theme.color.yellow};
+  padding: 1rem 1rem;
+
+  text-align: center;
+  border-radius: 1rem;
+
+  box-shadow: 0.2rem 0.2rem ${({ theme }) => theme.color.blue};
+  &.active {
+    background-color: ${({ theme }) => theme.color.blue};
+    box-shadow: 0.2rem 0.2rem ${({ theme }) => theme.color.yellow};
+  }
+`;
